+++ packages/examples/configs/pkg/index.js
@@ -1,0 +1,1 @@
+// do release script example here